requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
distro==1.9.0
eval_type_backport==0.2.2
fastapi==0.115.12
ffmpeg-python==0.2.0
future==1.0.0
google-auth==2.40.2
google-genai==1.16.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
jiter==0.10.0
mistralai==1.7.1
moviepy==2.2.1
numpy==2.2.6
openai==1.82.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
proglog==0.1.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
